[Driver] select alternative target containing . in executable name
The gcc compatible driver has support for selecting an alternative
target based on the driver's executable name, for instance
x86_64-unknown-linux-gnu-clang will set the target to linux on x86_64.

Previously, this failed when the target contains a minor version, for
example x86_64-unknown-freebsd13.1, so instead of finding the file's
stem, use the whole file name, but strip off any '.exe' from the tail.